Uninstall modules before upgrading files
